refactor: PairItem 파일 분리를 통한 가독성 개선
